% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsjs.R
\docType{package}
\name{ngsjs}
\alias{ngsjs}
\alias{ngsjs-package}
\title{Can be used to install all ngsjs command line tools required packages (CRAN, bioconductor and GitHub).}
\description{
Can be used to install all ngsjs command line tools required packages (CRAN, bioconductor and GitHub).
}
\seealso{
Useful links:

\url{https://github.com/ngsjs/ngsjs/src/ngsjs}

Report bugs at \url{https://github.com/ngsjs/ngsjs/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
