% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsjs.R
\docType{package}
\name{ngsjs}
\alias{ngsjs}
\alias{ngsjs-package}
\title{Provides R functions imported by command line tools in ngsjs.}
\description{
Provides R functions imported by command line tools in ngsjs.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/BioInstaller}

Report bugs at \url{https://github.com/JhuangLab/BioInstaller/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
